+++ static_test/geometry/transform/function.cpp
@@ -4,13 +4,13 @@
 namespace {
 static_assert([]() constexpr {
   constexpr auto transform =
-      usagi::geometry::transform<int>{usagi::geometry::point<int>{10, 20}, 0,
-                                      usagi::geometry::point<int>{2, 3},
-                                      usagi::geometry::point<int>{5, 7}};
-  constexpr auto local = usagi::geometry::point<int>{8, 9};
+      usagi::geometry::transform<float>{usagi::geometry::point<float>{10.f, 20.f}, 0.f,
+                                        usagi::geometry::point<float>{2.f, 3.f},
+                                        usagi::geometry::point<float>{5.f, 7.f}};
+  constexpr auto local = usagi::geometry::point<float>{8.f, 9.f};
   constexpr auto transformed = usagi::geometry::transform_point(transform, local);
 
-  if (transformed != usagi::geometry::point<int>{21, 33}) {
+  if (transformed != usagi::geometry::point<float>{21.f, 33.f}) {
     return false;
   }
